# [FEATURE] faster number conversion to String using Boost::Karma