--- a/src/openms/configh.cmake
+++ b/src/openms/configh.cmake
@@ -158,17 +158,6 @@ CHECK_INCLUDE_FILE_CXX("sys/time.h"  OPENMS_HAS_SYS_TIME_H)
 CHECK_INCLUDE_FILE_CXX("stdint.h"  OPENMS_HAS_STDINT_H)
 
 #------------------------------------------------------------------------------
-# check for libc++ bug
-try_run(_stream_bug_run_result_var _stream_bug_compile_var
-        ${CMAKE_BINARY_DIR}
-        ${OPENMS_HOST_DIRECTORY}/cmake/modules/check_string_stream_bug.cxx)
-
-# set stream variable
-if(NOT _stream_bug_run_result_var)
-  set(OPENMS_HAS_STREAM_EXTRACTION_BUG "1")
-endif()
-
-#------------------------------------------------------------------------------
 # check if certain c++ functions exist
 include(CheckFunctionExists)
 ## in MinGW we have the signal.h header, but no kill() as in Linux, so we need to check for the kill() function