--- a/src/openms/include/OpenMS/CONCEPT/PrecisionWrapper.h
+++ b/src/openms/include/OpenMS/CONCEPT/PrecisionWrapper.h
@@ -36,9 +36,8 @@
 
 #include <OpenMS/config.h>
 #include <OpenMS/CONCEPT/Types.h>
-
+#include <OpenMS/DATASTRUCTURES/String.h>
 #include <iostream>
-#include <iomanip>
 
 namespace OpenMS
 {
@@ -102,24 +101,11 @@ private:
   template <typename FloatingPointType>
   inline std::ostream & operator<<(std::ostream & os, const PrecisionWrapper<FloatingPointType> & rhs)
   {
-    // Same test as used by isnan(), spelled out here to avoid issues during overload resolution.
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wfloat-equal"
-    if (rhs.ref_ != rhs.ref_)
-#pragma clang diagnostic pop
-    {
-      // That's what Linux GCC uses, and gnuplot understands.
-      // Windows would print stuff like 1.#QNAN which makes testing hard.
-      return os << "nan";
-    }
-    else
-    {
-      const std::streamsize prec_save = os.precision();
-      os.precision(writtenDigits(FloatingPointType()));
-      os << rhs.ref_;
-      os.precision(prec_save);
-      return os;
-    }
+    // manual conversion to String is much faster than ostreams internal conversion
+    // (this calls the correct overload for extra precision)
+    String s(rhs.ref_, true);
+    os << s;
+    return os;
   }
 } // namespace OpenMS