--- a/src/openms/include/OpenMS/config.h.in
+++ b/src/openms/include/OpenMS/config.h.in
@@ -121,9 +121,6 @@
 
 #cmakedefine ENABLE_UPDATE_CHECK
 
-// is libc++ or has stream bug
-#cmakedefine OPENMS_HAS_STREAM_EXTRACTION_BUG
-
 // library versions
 #define OPENMS_LIBSVM_VERSION @CF_OPENMS_LIBSVM_VERSION@
 #define OPENMS_LIBSVM_VERSION_MAJOR @CF_OPENMS_LIBSVM_VERSION_MAJOR@