--- a/src/openms/source/ANALYSIS/OPENSWATH/SwathQC.cpp
+++ b/src/openms/source/ANALYSIS/OPENSWATH/SwathQC.cpp
@@ -43,6 +43,7 @@
 #include <nlohmann/json.hpp>
 
 #include <cmath>       // round
+#include <iomanip>     // setw
 
 namespace OpenSwath
 {