… cmake/modules/check_string_stream_bug.cxx
… src/openms/source/DATASTRUCTURES/DataValue.cpp
… src/openms/source/DATASTRUCTURES/String.cpp
… src/openms/source/FORMAT/IdXMLFile.cpp
… src/openms/source/FORMAT/PepNovoInfile.cpp
… src/openms/source/FORMAT/SequestInfile.cpp
… src/openms/source/SIMULATION/LABELING/SILACLabeler.cpp
… src/openms/source/SYSTEM/StopWatch.cpp
… src/pyOpenMS/tests/unittests/test000.py
… src/tests/class_tests/openms/data/AccurateMassSearchEngine_output1.consensusXML
… src/tests/class_tests/openms/data/AccurateMassSearchEngine_output1.featureXML
… src/tests/class_tests/openms/data/AccurateMassSearchEngine_output1_consensusXML.mzTab
… src/tests/class_tests/openms/data/ConsensusXMLFile_1.consensusXML
… src/tests/class_tests/openms/data/IdXMLFile_no_proteinhits.idXML
… src/tests/class_tests/openms/data/MascotInfile_test_template1.txt
… src/tests/class_tests/openms/data/MascotInfile_test_template2.txt
… src/tests/class_tests/openms/data/MascotInfile_test_template3.txt
… src/tests/class_tests/openms/data/PepNovoInfile_test_template.txt
… src/tests/class_tests/openms/data/SequestInfile_test_template1.txt
… src/tests/class_tests/openms/data/ToyExample1.traML
… src/tests/class_tests/openms/source/AASequence_test.cpp
… src/tests/class_tests/openms/source/ControlledVocabulary_test.cpp
… src/tests/class_tests/openms/source/DBoundingBox_test.cpp
… src/tests/class_tests/openms/source/DataFilters_test.cpp
… src/tests/class_tests/openms/source/DataValue_test.cpp
… src/tests/class_tests/openms/source/FuzzyStringComparator_test.cpp
… src/tests/class_tests/openms/source/ItraqConstants_test.cpp
… src/tests/class_tests/openms/source/LibSVMEncoder_test.cpp
… src/tests/class_tests/openms/source/MRMAssay_test.cpp
… src/tests/class_tests/openms/source/MRMDecoy_test.cpp