@hroest 2019-02-13
looks great, that should be a real improvement!

Can you look at these failed tests: https://travis-ci.org/OpenMS/OpenMS/jobs/492738641


```
The following tests FAILED:
	142 - ConsensusXMLFile_test (Failed)
	156 - IdXMLFile_test (Failed)
	160 - InspectInfile_test (Failed)
	201 - SequestInfile_test (Failed)
	211 - TraMLFile_test (Failed)
	1877 - TOPP_InspectAdapter_1 (Failed)
	1878 - TOPP_InspectAdapter_1_out1 (Failed)
	1879 - TOPP_InspectAdapter_2 (Failed)
	1880 - TOPP_InspectAdapter_2_out1 (Failed)
	2131 - TOPP_MaRaClusterAdapter_1_out_1 (Failed)
	2133 - TOPP_MaRaClusterAdapter_2_out_1 (Failed)
```