@cbielow 2019-02-14
fun fact:
calling `maracluster.exe -p 20ppm   ...` yields a slightly different result than `maracluster.exe -p 20.0ppm   ...`, even though the parameter is specified as double in their help section explicitly. And integers are perfectly representable as doubles (no rounding error)... go figure...