@cbielow 2019-02-15
more fun facts:
appleclang cannot parse `5.0e-06` as double using `stringstream >> d` (it will read 5.0 and then stop). All other compilers are just fine (and appleclang is also fine when reading `5.5e-06`.... 
see:
```
in1:  /Users/builder/jenkins/ws/openms_pr_ci/c797036f/build/src/tests/class_tests/openms/MascotXMLFile_test_249.tmp   (line: 646, position/column: 49/78)
				<UserParam type="float" name="EValue" value="!
				                                             5.0e-06"/>

in2:  /Users/builder/jenkins/ws/openms_pr_ci/c797036f/source/src/tests/class_tests/openms/data/MascotXMLFile_test_out_3.idXML   (line: 646, position/column: 49/78)
				<UserParam type="float" name="EValue" value="!
				                                             5e-06"/>
```