@cbielow 2019-02-20
ok, this is ready for merge.